Split out getEditedPost
